pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "classifier-runtime"
version = "0.1.0"
description = "DistilBERT text classification with ONNX Runtime CPU inference"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "transformers",
    "optimum[onnxruntime]",
    "onnx",
    "onnxruntime",
    "fastapi",
    "uvicorn",
    "pydantic",
    "numpy",
    "torch",
    "pyyaml",
    "scikit-learn",
    "tqdm",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx",
]

[project.scripts]
export-model = "classifier_runtime.models.classification.distilbert.export_model:main"
verify-parity = "classifier_runtime.models.classification.distilbert.verify_parity:main"
serve = "classifier_runtime.models.classification.distilbert.inference_service:run_server"

[tool.setuptools.packages.find]
include = ["classifier_runtime*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
